requirements.txt
torch==2.6.0
torchvision==0.21.0
numpy==2.2.3
pillow==10.4.0
tqdm==4.67.1
matplotlib==3.10.1
